refactor(mobile): keep session list screen focused
