+++ apps/mobile/src/components/agents/session-list-screen.tsx
@@ -11,28 +11,10 @@ import { AgentSessionListContent, FAB_MARGIN } from '@/components/agents/session
 import { SessionListHeaderActions } from '@/components/agents/session-list-header-actions';
 import { SessionListSearchHeader } from '@/components/agents/session-list-search-header';
 import { useSessionSearchInput } from '@/components/agents/use-session-search-input';
-import {
-  type ProjectFilterOption,
-  SessionFilterChips,
-  SessionFilterModal,
-} from '@/components/agents/platform-filter-modal';
-import {
-  excludeActiveFromGroups,
-  expandPlatformFilter,
-  formatGitUrlProject,
-  selectPinnedActiveSessions,
-  type SessionSection,
-} from '@/components/agents/session-list-helpers';
-import {
-  selectEffectiveSearchQuery,
-  selectShowSearchBusy,
-} from '@/components/agents/session-list-search-busy';
+import { SessionFilterChips, SessionFilterModal } from '@/components/agents/platform-filter-modal';
+import { selectShowSearchBusy } from '@/components/agents/session-list-search-busy';
+import { useAgentSessionListData } from '@/components/agents/use-agent-session-list-data';
 import { ScreenHeader } from '@/components/screen-header';
-import {
-  useAgentSessions,
-  useAgentSessionSearch,
-  useRecentAgentRepositories,
-} from '@/lib/hooks/use-agent-sessions';
 import { usePersistedAgentSessionFilters } from '@/lib/hooks/use-persisted-agent-session-filters';
 import { useOrganization } from '@/lib/organization-context';
 import { useThemeColors } from '@/lib/hooks/use-theme-colors';
@@ -47,12 +29,7 @@ export function AgentSessionListScreen() {
   const { fontScale } = useWindowDimensions();
 
   const tabBarHeight = useMemo(
-    () =>
-      getEffectiveTabBarHeight({
-        bottomInset: bottom,
-        platform: Platform.OS,
-        fontScale,
-      }),
+    () => getEffectiveTabBarHeight({ bottomInset: bottom, platform: Platform.OS, fontScale }),
     [bottom, fontScale]
   );
 
@@ -79,100 +56,40 @@ export function AgentSessionListScreen() {
     searchController,
   } = useSessionSearchInput();
 
-  const createdOnPlatform = useMemo(
-    () => (platformFilter.length > 0 ? expandPlatformFilter(platformFilter) : undefined),
-    [platformFilter]
-  );
-  const gitUrl = useMemo(
-    () => (projectFilter.length > 0 ? projectFilter : undefined),
-    [projectFilter]
-  );
-
   const ready = filtersLoaded && orgLoaded;
+
+  // Navigation guard: prevent double-push on rapid row taps. Re-arms on next focus.
+  const rowNavLockRef = useRef(false);
+
   const {
     storedSessions,
-    dateGroups,
     activeSessions,
-    activeSessionIds,
     activeIsError,
     isLoading,
-    storedIsError,
     hasNextPage,
     isFetchingNextPage,
     fetchNextPage,
     refetch,
-  } = useAgentSessions({
-    createdOnPlatform,
-    gitUrl,
-    organizationId,
-    enabled: ready,
-    sortBy,
-  });
-  const isSearching = searchQuery.length > 0;
-  const search = useAgentSessionSearch({
-    searchQuery,
-    createdOnPlatform,
-    gitUrl,
-    organizationId,
-    enabled: ready && isSearching,
-    sortBy,
-  });
-  const showSearchBusy = selectShowSearchBusy({
-    awaitingCommit,
+    handleRetry,
+    handleRefetch,
     isSearching,
-    isFetching: search.isFetching,
-  });
-  const { data: recentRepositories } = useRecentAgentRepositories({
+    search,
+    projectOptions,
+    contentIsError,
+    pinnedActive,
+    sections,
+  } = useAgentSessionListData({
     organizationId,
-    enabled: ready,
+    platformFilter,
+    projectFilter,
+    sortBy,
+    ready,
+    searchQuery,
   });
-
-  // The body's error/empty state is driven only by the query that actually
-  // fills the body: the search query while searching, otherwise the stored
-  // (history) list. A transient active-poll blip must never fold into this —
-  // it surfaces solely through the inline "Couldn't refresh" line via
-  // `activeIsError`. Retrying still hits whichever query is really in error
-  // instead of always refetching the base list underneath a failed search;
-  // an active-only failure during search additionally retries the active poll
-  // so the inline staleness line clears.
-  const contentIsError = isSearching ? search.isError : storedIsError;
-
-  const searchRefetch = search.refetch;
-  const handleRetry = useCallback(() => {
-    if (!isSearching) {
-      void refetch();
-      return;
-    }
-    if (activeIsError) {
-      // search is the body, active is the tray — retry both.
-      void (async () => {
-        await Promise.all([searchRefetch(), refetch()]);
-      })();
-      return;
-    }
-    void searchRefetch();
-  }, [activeIsError, isSearching, refetch, searchRefetch]);
-
-  // Pull-to-refresh must also retry the search query while one is active —
-  // it's the query actually driving what's on screen.
-  const handleRefetch = useCallback(async () => {
-    if (!isSearching) {
-      await refetch();
-      return;
-    }
-    await Promise.all([searchRefetch(), refetch()]);
-  }, [isSearching, refetch, searchRefetch]);
-
   const refetchRef = useRef(refetch);
   useEffect(() => {
     refetchRef.current = refetch;
   }, [refetch]);
-
-  // Navigation guard: prevent double-push on rapid row taps. One lock covers
-  // history + tray presses (both funnel through `navigateToSession`). Re-arms
-  // on next focus so returning from the detail screen always releases the guard.
-  const rowNavLockRef = useRef(false);
-
   useFocusEffect(
     useCallback(() => {
       rowNavLockRef.current = false;
@@ -180,54 +97,12 @@ export function AgentSessionListScreen() {
     }, [])
   );
 
-  const projectOptions = useMemo((): ProjectFilterOption[] => {
-    const byGitUrl = new Map<string, ProjectFilterOption>();
-
-    const repositories = recentRepositories?.repositories ?? [];
-    for (const project of repositories.slice(0, 3)) {
-      byGitUrl.set(project.gitUrl, {
-        gitUrl: project.gitUrl,
-        displayName: formatGitUrlProject(project.gitUrl),
-      });
-    }
-
-    for (const selectedGitUrl of projectFilter) {
-      if (!byGitUrl.has(selectedGitUrl)) {
-        byGitUrl.set(selectedGitUrl, {
-          gitUrl: selectedGitUrl,
-          displayName: formatGitUrlProject(selectedGitUrl),
-        });
-      }
-    }
-
-    return [...byGitUrl.values()];
-  }, [projectFilter, recentRepositories?.repositories]);
-
-  // While the first fetch for this search text is still in flight (no
-  // keepPreviousData to fall back on yet), render as if no search were
-  // applied instead of blanking to an empty/mismatched list —
-  // `showSearchBusy` drives a lightweight inline indicator instead.
-  const effectiveSearchQuery = selectEffectiveSearchQuery({
+  const showSearchBusy = selectShowSearchBusy({
+    awaitingCommit,
     isSearching,
-    isPending: search.isPending,
-    searchQuery,
+    isFetching: search.isFetching,
   });
 
-  // Pinned "Active now" tray. The committed effective search query narrows
-  // the tray together with the history data source (blank while the first
-  // fetch for the text is pending), in conjunction with the platform/project
-  // filters, so the tray never shows a session the user has searched or
-  // filtered away.
-  const pinnedActive = useMemo(
-    () =>
-      selectPinnedActiveSessions({
-        activeSessions,
-        projectFilter,
-        platformFilter,
-        searchQuery: effectiveSearchQuery,
-      }),
-    [activeSessions, effectiveSearchQuery, platformFilter, projectFilter]
-  );
   const hasPinnedActive = pinnedActive.length > 0;
 
   const organizationIdBySessionId = useMemo(
@@ -235,31 +110,19 @@ export function AgentSessionListScreen() {
     [storedSessions]
   );
 
-  // History sections only. The pinned tray takes over for active sessions
-  // and `excludeActiveFromGroups` keeps history exclusivity.
-  const sections = useMemo<SessionSection[]>(() => {
-    // Stored sessions are cursor-paginated, so a client-side filter would only
-    // see the loaded pages. When a query is active, use the server search
-    // results (which cover the full history) instead.
-    const storedGroups = effectiveSearchQuery ? search.dateGroups : dateGroups;
-    return excludeActiveFromGroups(storedGroups, activeSessionIds).map(group => ({
-      title: group.label,
-      data: group.sessions,
-    }));
-  }, [activeSessionIds, dateGroups, effectiveSearchQuery, search.dateGroups]);
-
   const navigateToSession = useCallback(
     (sessionId: string, sessionOrgId?: string | null) => {
-      if (rowNavLockRef.current) return;
+      if (rowNavLockRef.current) {
+        return;
+      }
       rowNavLockRef.current = true;
       try {
-        const path = sessionOrgId
-          ? `/(app)/agent-chat/${sessionId}?organizationId=${sessionOrgId}`
-          : `/(app)/agent-chat/${sessionId}`;
-        router.push(path as Href);
+        router.push(
+          (sessionOrgId
+            ? `/(app)/agent-chat/${sessionId}?organizationId=${sessionOrgId}`
+            : `/(app)/agent-chat/${sessionId}`) as Href
+        );
       } catch {
-        // If push throws synchronously (e.g. duplicate route), release the
-        // lock so the row stays tappable.
         rowNavLockRef.current = false;
       }
     },
@@ -275,12 +138,8 @@ export function AgentSessionListScreen() {
   const hasActiveFilter = platformFilter.length > 0 || projectFilter.length > 0;
   const hasAnySessions = storedSessions.length > 0 || activeSessions.length > 0;
 
-  // Search header is rendered at the screen level (above the scrolling list)
-  // so it stays reachable without scrolling. The Active now tray scrolls
-  // inside the list header. Recompute the body's `showInlineError` here
-  // via the SAME pure selector, with the same inputs, so the inline
-  // "Couldn't refresh" line stays identical and the body-model test keeps
-  // covering it.
+  // Inline error recomputed here (same pure selector, same inputs) so the
+  // body-model test continues covering it.
   const showInlineError = useMemo(
     () =>
       selectSessionListBodyModel({
@@ -298,6 +157,11 @@ export function AgentSessionListScreen() {
     clearSearchInput();
     searchController.clearBroadly(setFilters);
   }, [clearSearchInput, searchController, setFilters]);
+
+  const fabStyle = useMemo(
+    () => ({ bottom: tabBarHeight + FAB_MARGIN, right: 20 }),
+    [tabBarHeight]
+  );
 
   return (
     <View className="flex-1 bg-background">
@@ -386,8 +250,7 @@ export function AgentSessionListScreen() {
           }}
         />
       )}
-      {/* FAB visible when there are sessions — empty state already owns
-          the creation CTA. Header "+" stays unchanged. */}
+      {/* FAB visible when there are sessions — empty state already owns the creation CTA. */}
       {hasAnySessions && (
         <Pressable
           accessibilityRole="button"
@@ -397,10 +260,7 @@ export function AgentSessionListScreen() {
             router.push(getNewAgentSessionPath(organizationId) as Href);
           }}
           className="absolute h-14 w-14 items-center justify-center rounded-full bg-primary shadow-lg shadow-black/25 active:opacity-80"
-          style={{
-            bottom: tabBarHeight + FAB_MARGIN,
-            right: 20,
-          }}
+          style={fabStyle}
         >
           <Plus size={24} color={colors.primaryForeground} />
         </Pressable>
